Refactor: Move utility functions that can be reused to a separate module
